requirements.txt
Pillow==10.3.0
protobuf==5.27.0
python-dotenv==1.0.1
streamlit==1.33.0
google-generativeai==0.5.0
